Fixed CUDA issues on driver and executor sides
Stop loading CUDAManager at driver if it is not local since a machine for driver may not have GPUs.
Stop storing CUDAManager into SparkContext since this will not be passed into executors.
Remove APIs in CUDAManager to register a CUDAKernel as a symbol since registered info will not be passed to executors.